1. Started writing tests for nano-blockchain crate. 2. Moved TemporaryBlockProducer struct used in blockchain-albatross crate tests to block-production-albatross crate.
