+++ blockchain-albatross/src/history_store/history_tree_proof.rs
@@ -1,8 +1,9 @@
+use merkle_mountain_range::mmr::proof::Proof;
+
 use beserial::{
     Deserialize, DeserializeWithLength, ReadBytesExt, Serialize, SerializeWithLength,
     SerializingError, WriteBytesExt,
 };
-use merkle_mountain_range::mmr::proof::Proof;
 use nimiq_hash::Blake2bHash;
 
 use crate::history_store::{ExtendedTransaction, HistoryTreeHash};
