+++ nano-blockchain/Cargo.toml
@@ -14,10 +14,15 @@ thiserror = "1.0"
 nimiq-account = { path = "../primitives/account" }
 nimiq-block-albatross = { path = "../primitives/block-albatross" }
 nimiq-blockchain-albatross = { path = "../blockchain-albatross" }
-nimiq-database = { path = "../database", features = ["full-nimiq"] }
+nimiq-database = { path = "../database" }
 nimiq-genesis = { path = "../genesis" }
 nimiq-hash = { path = "../hash" }
 nimiq-nano-sync = { path = "../nano-sync", default-features = false, features = ["verifier"]  }
 nimiq-primitives = { path = "../primitives", features = ["policy"] }
 nimiq-tree-primitives = { path = "../accounts/tree-primitives" }
-nimiq-utils = { path = "../utils", features = ["merkle", "time"] }
+nimiq-utils = { path = "../utils", features = ["time"] }
+
+[dev-dependencies]
+rand = "0.8.0"
+
+nimiq-block-production-albatross = { path = "../block-production-albatross", features = ["test-utils"] }
